# openapi/spec.yml
components:
  schemas:
...
    permissionCreate:
      properties:
        action:
          type: string
        entity:
          type: string
        id:
          type: string
        name:
          type: string
      required:
      - id
      - name
      - entity
      - action
      type: object
    permissionPartial:
      properties:
        action:
          type: string
        entity:
          type: string
        id:
          type: string
        name:
          type: string
      type: object
...
    permissionUpdate:
      properties:
        action:
          type: string
        entity:
          type: string
        id:
          type: string
        name:
          type: string
      required:
      - id
      - name
      - entity
      - action
      type: object
    permissionView:
      properties:
        action:
          type: string
        entity:
          type: string
        id:
          type: string
        name:
          type: string
        roles:
          items:
            $ref: '#/components/schemas/rolePermissionView'
          type: array
      type: object
...
    roleCreate:
      properties:
        id:
          type: string
        name:
          type: string
      required:
      - id
      - name
      type: object
    rolePartial:
      properties:
        id:
          type: string
        name:
          type: string
      type: object
...
    rolePermissionView:
      properties:
        id:
          format: uuid
          type: string
        permission:
          $ref: '#/components/schemas/permissionView'
        permission_id:
          type: string
        role:
          $ref: '#/components/schemas/roleView'
        role_id:
          type: string
      type: object
...
    roleUpdate:
      properties:
        id:
          type: string
        name:
          type: string
      required:
      - id
      - name
      type: object
    roleView:
      properties:
        id:
          type: string
        name:
          type: string
        rolePermissions:
          items:
            $ref: '#/components/schemas/rolePermissionView'
          type: array
      type: object
...
  securitySchemes:
    bearer:
      scheme: bearer
      type: http
    cookieAccessToken:
      in: cookie
      name: x-access-token
      type: apiKey
    cookieRefreshToken:
      in: cookie
      name: x-refresh-token
      type: apiKey
    key:
      in: header
      name: X-API-Key
      type: apiKey
    oauth:
      flows:
        authorizationCode:
          authorizationUrl: https://example.com/oauth/authorize
          scopes:
            admin: Grants access to admin operations
            read: Grants read access
            write: Grants write access
          tokenUrl: https://example.com/oauth/token
      type: oauth2
info:
  description: Invoicing API Reference
  title: Invoicing API
  version: '1.0'
openapi: 3.0.3
paths:
...
  /api/v1/permissions:
    get:
      description: get all permission entities
      operationId: getPermissions
      parameters:
      - description: Filter Permissions by name
        in: query
        name: name
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/permissionView'
                type: array
          description: successful operation
      summary: get all permission entities
      tags:
      - permission
    post:
      description: Add a new permission entity
      operationId: addPermission
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/permissionCreate'
        description: Create a new permission entity
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/permissionView'
          description: successful operation
        '405':
          description: invalid input
      summary: Add a new permission entity
      tags:
      - permission
  /api/v1/permissions/{id}:
    delete:
      description: Delete an existing permission entity
      operationId: deletePermission
      parameters:
      - description: permission id to delete
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/permissionView'
          description: successful operation
        '400':
          description: Invalid permission id
        '404':
          description: permission not found
      summary: Delete an existing permission entity
      tags:
      - permission
    get:
      description: Get an existing permission entity
      operationId: getPermission
      parameters:
      - description: permission id to get
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/permissionView'
          description: successful operation
        '400':
          description: Invalid permission id
        '404':
          description: permission not found
      summary: Get an existing permission entity
      tags:
      - permission
    patch:
      description: Modify an existing permission entity (partial update, only provided
        properties will be updated)
      operationId: modifyPermission
      parameters:
      - description: permission id to partially update
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/permissionPartial'
        description: Modify an existing permission entity
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/permissionView'
          description: successful operation
        '400':
          description: Invalid ID
        '404':
          description: permission not found
        '405':
          description: Invalid input
      summary: Modify an existing permission entity
      tags:
      - permission
    put:
      description: Update an existing permission entity (all required properties must
        be supplied)
      operationId: updatePermission
      parameters:
      - description: permission id to update
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/permissionUpdate'
        description: Update an existing permission entity
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/permissionView'
          description: successful operation
        '400':
          description: Invalid ID
        '404':
          description: permission not found
        '405':
          description: Invalid input
      summary: Update an existing permission entity
      tags:
      - permission
...
  /api/v1/roles:
    get:
      description: get all role entities
      operationId: getRoles
      parameters:
      - description: Filter Roles by name
        in: query
        name: name
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/roleView'
                type: array
          description: successful operation
      summary: get all role entities
      tags:
      - role
    post:
      description: Add a new role entity
      operationId: addRole
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/roleCreate'
        description: Create a new role entity
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/roleView'
          description: successful operation
        '405':
          description: invalid input
      summary: Add a new role entity
      tags:
      - role
  /api/v1/roles/{id}:
    delete:
      description: Delete an existing role entity
      operationId: deleteRole
      parameters:
      - description: role id to delete
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/roleView'
          description: successful operation
        '400':
          description: Invalid role id
        '404':
          description: role not found
      summary: Delete an existing role entity
      tags:
      - role
    get:
      description: Get an existing role entity
      operationId: getRole
      parameters:
      - description: role id to get
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/roleView'
          description: successful operation
        '400':
          description: Invalid role id
        '404':
          description: role not found
      summary: Get an existing role entity
      tags:
      - role
    patch:
      description: Modify an existing role entity (partial update, only provided properties
        will be updated)
      operationId: modifyRole
      parameters:
      - description: role id to partially update
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/rolePartial'
        description: Modify an existing role entity
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/roleView'
          description: successful operation
        '400':
          description: Invalid ID
        '404':
          description: role not found
        '405':
          description: Invalid input
      summary: Modify an existing role entity
      tags:
      - role
    put:
      description: Update an existing role entity (all required properties must be
        supplied)
      operationId: updateRole
      parameters:
      - description: role id to update
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/roleUpdate'
        description: Update an existing role entity
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/roleView'
          description: successful operation
        '400':
          description: Invalid ID
        '404':
          description: role not found
        '405':
          description: Invalid input
      summary: Update an existing role entity
      tags:
      - role
...
security:
- cookieAccessToken: []
  cookieRefreshToken: []
servers:
- url: http://localhost:8001
...